Rename module LiskClient -> WsApi
