+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from "./HttpApi";
-export * from "./LiskClient";
 export * from "./LiskPeer";
+export * from "./WsApi";
